decouple SSR module execution from Workers implementation
